main: Use --limit-files
Earlier --files was being used but that wasn't the correct
option to be used because that replaces the files config value
only in [Default] section. This means that any other
section would clash with it.
Now we use the --limit-files argument which takes a subset
of all files with the glob provided in --limit-files. This ensures
that all sections with this file get executed and no clashes happen.